add endpoints & bug fixes

+++ src/main/java/com/example/demo/Service/DepartmentService.java
@@ -1,6 +1,7 @@
 package com.example.demo.Service;
 
 import com.example.demo.Model.DAO.Department;
+import com.example.demo.Model.DAO.Employee;
 import com.example.demo.Repository.DepartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -29,6 +30,18 @@ public class DepartmentService {
         return department;
     }
     public void deleteDepartmentById(int id){
-        this.departmentRepository.deleteById(id);
+        EmployeeService employeeService = new EmployeeService();
+        List<Employee> employeeList = employeeService.findAllEmployees();
+        int flag=0;
+        for(Employee employee:employeeList){
+            Department department = new Department();
+            department=employee.getDepartment();
+            if(department.getId()==id){
+                flag=1;
+            }
+        }
+        if(flag == 0){
+            this.departmentRepository.deleteById(id);
+        }
     }
 }
